trip_planner/pyproject.toml
[project]
name = "trip-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "reportlab>=4.2.0",
    "requests>=2.32.0",
    "pypdf>=4.0.0",
    "whitenoise>=6.11.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "gunicorn>=23.0.0",
    "drf-spectacular>=0.29.0",
    "djangorestframework-simplejwt>=5.5.1",
]
